refactor: add update mixin
