+++ docarray/base_document/mixins/__init__.py
@@ -1,3 +1,4 @@
 from docarray.base_document.mixins.proto import ProtoMixin
+from docarray.base_document.mixins.update import UpdateMixin
 
-__all__ = ['ProtoMixin']
+__all__ = ['ProtoMixin', 'UpdateMixin']
